MEP22 Save Figure Tool
