+++ lib/matplotlib/backend_tools.py
@@ -19,6 +19,8 @@ import six
 import time
 import warnings
 
+import os
+
 
 class Cursors(object):
     """Simple namespace for cursor reference"""
@@ -677,6 +679,43 @@ class SaveFigureBase(ToolBase):
     description = 'Save the figure'
     image = 'filesave.png'
     default_keymap = rcParams['keymap.save']
+
+
+class ToolSaveFigure(ToolBase):
+    """Saves the figure"""
+
+    description = 'Save the figure'
+    image = 'filesave.png'
+    default_keymap = rcParams['keymap.save']
+
+    def get_filechooser(self):
+        fc = self.figure.canvas.backend.FileChooserDialog(
+            title='Save the figure',
+            parent=self.figure.canvas.manager.window,
+            path=os.path.expanduser(rcParams.get('savefig.directory', '')),
+            filetypes=self.figure.canvas.get_supported_filetypes(),
+            default_filetype=self.figure.canvas.get_default_filetype())
+        fc.set_current_name(self.figure.canvas.get_default_filename())
+        return fc
+
+    def trigger(self, *args, **kwargs):
+        chooser = self.get_filechooser()
+        fname, format_ = chooser.get_filename_from_user()
+        chooser.destroy()
+        if fname:
+            startpath = os.path.expanduser(
+                rcParams.get('savefig.directory', ''))
+            if startpath == '':
+                # explicitly missing key or empty str signals to use cwd
+                rcParams['savefig.directory'] = startpath
+            else:
+                # save dir for next time
+                rcParams['savefig.directory'] = os.path.dirname(
+                    six.text_type(fname))
+            try:
+                self.figure.canvas.print_figure(fname, format=format_)
+            except Exception as e:
+                error_msg_gtk(str(e), parent=self)
 
 
 class ZoomPanBase(ToolToggleBase):
