+++ lib/matplotlib/backends/backend_gtk3agg.py
@@ -124,6 +124,8 @@ FigureManager = FigureManagerGTK3Agg
 Window = backend_gtk3.WindowGTK3
 Toolbar = backend_gtk3.ToolbarGTK3
 Statusbar = backend_gtk3.StatusbarGTK3
+FileChooserDialog = backend_gtk3.FileChooserDialog
+
 Toolbar2 = backend_gtk3.NavigationToolbar2GTK3
 MainLoop = backend_gtk3.MainLoopGTK3
 show = backend_gtk3.show
